Vuelco del portatil con varios cambios 2
- Cambio catálogo para que se actualice con la base de datos.
- Modifico Script base de datos.
- Modifico varios html y css tablas.
- Creo página contacto.html
- Creo login2.html para redirigir restringiendo acceso.
- Implemento la recogida de datos de sesión en ServletLogin.
- Implemento los permisos de sesión en los servlets:
	*Administrador:
	- ServletListarUsuarios
	- ServletListarReservas
	- ServletListarMotos
	- ServletEliminarMoto
	- ServletEliminarReserva
	- ServletEliminarUsuario
	- ServletModificarMoto
	- ServletModificarReserva ----> TENGO QUE TERMINAR ESTE SERVLET
	- ServletModificarUsuario
- Modifico adminArea.html y clientArea.html para cerrar sesión.
- Modifico adminArea.css
- Modifico clientRes y clientUsu.html
- Creo página sesionCerrada.html
- Cambio ServletClienteLeerReservas, Reserva y DaoReserva para mostrar en la sesión del cliente las reservas realizadas CON SU ID.
- Creo ServletCatalogo para mostrar los datos de los productos sin restricciones de usuario.
- Modifico el formulario de insertarReserva y el servlet para que recoja el idSesion de la sesion y grabe la reserba con él.
- Creo condic.html.

+++ Alca-Ride Proyect/src/main/java/controlador/ServletEliminarUsuario.java
@@ -5,6 +5,7 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 import modelo.Login;
 import modelo.Usuario;
 
@@ -19,6 +20,8 @@ import DAO.DaoUsuario;
  */
 public class ServletEliminarUsuario extends HttpServlet {
 	private static final long serialVersionUID = 1L;
+	//Inicializo el objeto Httpsession y la variable como atributo del servlet que voy a utilizar.
+	HttpSession sesion;
        
     /**
      * @see HttpServlet#HttpServlet()
@@ -40,47 +43,59 @@ public class ServletEliminarUsuario extends HttpServlet {
 	 * @see HttpServlet#doPost(HttpServletRequest request, HttpServletResponse response)
 	 */
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        // Se obtiene el ID del usuario que se va a borrar desde los parámetros de la solicitud
-        int idParaBorrar = Integer.parseInt(request.getParameter("id_Usuario"));
+		        
+		//Control de sesión para ADMINISTRADOR
+		HttpSession sesion = request.getSession();
+									
+		//Si tengo los datos que necesito, compruebo si son de administrador o de cliente.
+		//Si son de administrador le enseño la tabla.
+		int idSesion = (Integer) sesion.getAttribute("id");
+		boolean isAdmin = (boolean) sesion.getAttribute("admin");
 
-        // Eliminar usuario de la tabla 'login'
-        try {
-            Login login = new Login(); // Se crea un objeto Login para representar al usuario
-            login.setId_Usuario(idParaBorrar); // Se establece el ID del usuario a borrar en el objeto Login
-            
-            login.eliminarLogin();
-            System.out.println("Usuario borrado correctamente de la tabla login"); // Se imprime un mensaje de éxito en la consola
-            
-            } catch (SQLException e) {
-            System.out.println("Error al eliminar el usuario de la tabla login: " + e.getMessage()); // Se imprime un mensaje de error en la consola si ocurre una excepción SQL
-            response.setContentType("text/plain");
-            response.setCharacterEncoding("UTF-8");
-            response.getWriter().write("Error al eliminar el login de la base de datos");
-            response.sendRedirect("adminUsu.html?mensaje=Error+al+eliminar+el+login+de+la+base+de+datos.");
-        }
-
-        // Eliminar usuario de la tabla 'usuarios'
-        try {
-            Usuario usuario = new Usuario(); // Se crea un objeto Usuario para representar al usuario
-            usuario.setId_Usuario(idParaBorrar); // Se establece el ID del usuario a borrar en el objeto Usuario
-            
-            usuario.eliminarUsuario();
-            System.out.println("Usuario borrado correctamente de la tabla usuarios"); // Se imprime un mensaje de éxito en la consola
-            
-            //Mando mensaje por ventana emergente a JavaScript
-            response.setContentType("text/plain");
-            response.setCharacterEncoding("UTF-8");
-            response.getWriter().write("Usuario eliminado correctamente");
-            response.sendRedirect("adminUsu.html?mensaje=Usuario+eliminado+correctamente");// Se redirige al usuario a la página 'adminUsu.html' después de borrar el usuario y se muestra mensaje de confirmación
-            
-        } catch (SQLException e) {
-            System.out.println("Error al eliminar el usuario de la tabla usuarios: " + e.getMessage()); // Se imprime un mensaje de error en la consola si ocurre una excepción SQL
-            response.setContentType("text/plain");
-            response.setCharacterEncoding("UTF-8");
-            response.getWriter().write("Error al eliminar el usuario de la base de datos");
-            response.sendRedirect("adminUsu.html?mensaje=Error+al+eliminar+el+usuario+de+la+base+de+datos.");
-        }
-    }
+		if (idSesion != 0 && isAdmin == true) {
+		
+				// Se obtiene el ID del usuario que se va a borrar desde los parámetros de la solicitud
+		        int idParaBorrar = Integer.parseInt(request.getParameter("id_Usuario"));
+		
+		        // Eliminar usuario de la tabla 'login'
+		        try {
+		            Login login = new Login(); // Se crea un objeto Login para representar al usuario
+		            login.setId_Usuario(idParaBorrar); // Se establece el ID del usuario a borrar en el objeto Login
+		            
+		            login.eliminarLogin();
+		            System.out.println("Usuario borrado correctamente de la tabla login"); // Se imprime un mensaje de éxito en la consola
+		            
+		            } catch (SQLException e) {
+		            System.out.println("Error al eliminar el usuario de la tabla login: " + e.getMessage()); // Se imprime un mensaje de error en la consola si ocurre una excepción SQL
+		            response.setContentType("text/plain");
+		            response.setCharacterEncoding("UTF-8");
+		            response.getWriter().write("Error al eliminar el login de la base de datos");
+		            response.sendRedirect("adminUsu.html?mensaje=Error+al+eliminar+el+login+de+la+base+de+datos.");
+		        }
+		
+		        // Eliminar usuario de la tabla 'usuarios'
+		        try {
+		            Usuario usuario = new Usuario(); // Se crea un objeto Usuario para representar al usuario
+		            usuario.setId_Usuario(idParaBorrar); // Se establece el ID del usuario a borrar en el objeto Usuario
+		            
+		            usuario.eliminarUsuario();
+		            System.out.println("Usuario borrado correctamente de la tabla usuarios"); // Se imprime un mensaje de éxito en la consola
+		            
+		            //Mando mensaje por ventana emergente a JavaScript
+		            response.setContentType("text/plain");
+		            response.setCharacterEncoding("UTF-8");
+		            response.getWriter().write("Usuario eliminado correctamente");
+		            response.sendRedirect("adminUsu.html?mensaje=Usuario+eliminado+correctamente");// Se redirige al usuario a la página 'adminUsu.html' después de borrar el usuario y se muestra mensaje de confirmación
+		            
+		        } catch (SQLException e) {
+		            System.out.println("Error al eliminar el usuario de la tabla usuarios: " + e.getMessage()); // Se imprime un mensaje de error en la consola si ocurre una excepción SQL
+		            response.setContentType("text/plain");
+		            response.setCharacterEncoding("UTF-8");
+		            response.getWriter().write("Error al eliminar el usuario de la base de datos");
+		            response.sendRedirect("adminUsu.html?mensaje=Error+al+eliminar+el+usuario+de+la+base+de+datos.");
+		        }
+		    }
+	}
 }
